feat: add logic to save created teams locally

+++ components/TeamCard.tsx
@@ -2,16 +2,27 @@
 import { Team } from "@/types/team"
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import { useState } from "react"
+import { useEffect, useState } from "react"
 
 export function TeamCard({team}: {team: Team}){
     const [isOpen, setIsOpen] = useState(false);
     const [pokemonName, setPokemonName] = useState("");
-
+    const [isMine, setIsMine] = useState(false);
     const router = useRouter();
+
+    useEffect(() => {
+        const savedIds = JSON.parse(localStorage.getItem('my_teams') || '[]');
+        if(savedIds.includes(team.id)){
+            setIsMine(true);
+        }
+    }, [team.id])
 
     async function addPoke(){
         if(!pokemonName){
+            return;
+        }
+        if(isMine == false){
+            alert("Você não pode adicionar pokemon ao time de outra pessoa!!");
             return;
         }
 
@@ -38,6 +49,10 @@ export function TeamCard({team}: {team: Team}){
         }
     }
     async function removeTeam(id : string){
+        if(isMine == false){
+            alert("Você não pode deletar o time de outra pessoa!!");
+            return;
+        }
         try{
             const res = await fetch(`http://localhost:3000/teams/${id}`,{
                 method: "DELETE",
@@ -55,6 +70,10 @@ export function TeamCard({team}: {team: Team}){
     }
 
     async function removePokemon(id: string, name:string){
+         if(isMine == false){
+            alert("Você não pode remover pokemon do time de outra pessoa!!");
+            return;
+        }
         try {
             const res = await fetch(`http://localhost:3000/teams/${id}/pokemon/${name}`, {
             method: "DELETE"
@@ -73,7 +92,16 @@ export function TeamCard({team}: {team: Team}){
     return(
         <>
         <div className="flex flex-col items-center justify-center gap-5 py-5 border rounded-lg shadow-lg hover:scale-105 bg-white">
-
+            {isMine && (
+                <span className="  bg-blue-600 text-white text-xs px-2 py-1 rounded-full">
+                    MEU
+                </span>
+            )}
+            {!isMine && (
+                <span className="  bg-white text-black border border-black text-xs px-2 py-1 rounded-full">
+                    OUTRO
+                </span>
+            )}
             <Link href={`/teams/${team.id}`} className="bg-black text-white rounded p-2">Detalhes do time</Link>
             <div>
                 <h2><strong>{team.name} </strong></h2>
@@ -85,14 +113,20 @@ export function TeamCard({team}: {team: Team}){
                     {team.pokemons.map((pokemon) => (
                     <li key={pokemon.id}>
                         {pokemon.name} 
-                        <button className="bg-red-500 rounded m-2 px-1 cursor-pointer" onClick={() => removePokemon(team.id, pokemon.name)}>Deletar</button>
-                    </li>
+                        {isMine && (
+                            <button className="bg-red-500 rounded m-2 px-1 cursor-pointer" onClick={() => removePokemon(team.id, pokemon.name)}>Deletar</button>
+                        )}
+                        </li>
                 ))}
                 </ul>
             </div>
-            <button onClick={() => setIsOpen(true)} className="bg-blue-500 h-10 cursor-pointer rounded px-2">Adicionar pokemon</button>
-            <button onClick={() => removeTeam(team.id)} className="bg-red-500 cursor-pointer rounded p-2">Deletar time</button>
-            
+             {isMine && (
+                <div className="flex flex-col gap-5 ">
+                    <button onClick={() => setIsOpen(true)} className="bg-blue-500 h-10 cursor-pointer rounded px-2">Adicionar pokemon</button>
+                    <button onClick={() => removeTeam(team.id)} className="bg-red-500 cursor-pointer rounded p-2">Deletar time</button>
+                </div>
+                
+            )}
         </div>
 
         {isOpen && (
